+++ esp32c3-embassy/src/main.rs
@@ -29,10 +29,9 @@ use embassy_sync::channel::Sender;
 use esp_alloc::heap_allocator;
 
 use esp_hal::clock::CpuClock;
-use esp_hal::dma::Dma;
 use esp_hal::dma::DmaBufError;
+use esp_hal::dma::DmaChannel0;
 use esp_hal::dma::DmaDescriptor;
-use esp_hal::dma::DmaPriority;
 use esp_hal::dma::DmaRxBuf;
 use esp_hal::dma::DmaTxBuf;
 use esp_hal::gpio::GpioPin;
@@ -41,27 +40,29 @@ use esp_hal::gpio::Level;
 use esp_hal::gpio::Output;
 use esp_hal::gpio::Pull;
 use esp_hal::i2c::master::Config as I2cConfig;
+use esp_hal::i2c::master::ConfigError as I2cConfigError;
 use esp_hal::i2c::master::I2c;
 use esp_hal::init as initialize_esp_hal;
-use esp_hal::peripherals::DMA;
 use esp_hal::peripherals::I2C0;
 use esp_hal::peripherals::RADIO_CLK;
 use esp_hal::peripherals::SPI2;
 use esp_hal::peripherals::TIMG0;
 use esp_hal::peripherals::WIFI;
-use esp_hal::prelude::*; // RateExtU32, main, ram
+use esp_hal::ram;
 use esp_hal::rng::Rng;
 use esp_hal::spi::master::Config as SpiConfig;
+use esp_hal::spi::master::ConfigError as SpiConfigError;
 use esp_hal::spi::master::Spi;
 use esp_hal::spi::master::SpiDma;
-use esp_hal::spi::SpiMode;
-use esp_hal::timer::systimer::SystemTimer;
-use esp_hal::timer::systimer::Target;
+use esp_hal::spi::Mode as SpiMode;
+use esp_hal::time::RateExtU32 as _;
 use esp_hal::timer::timg::TimerGroup;
 use esp_hal::Async;
+use esp_hal::Blocking;
 use esp_hal::Config as EspConfig;
 
 use esp_hal_embassy::init as initialize_embassy;
+use esp_hal_embassy::main;
 
 use time::OffsetDateTime;
 
@@ -205,8 +206,8 @@ async fn main_fallible(
 
     heap_allocator!(HEAP_MEMORY_SIZE);
 
-    let systimer = SystemTimer::new(peripherals.SYSTIMER).split::<Target>();
-    initialize_embassy(systimer.alarm0);
+    let timg1 = TimerGroup::new(peripherals.TIMG1);
+    initialize_embassy(timg1.timer0);
 
     let rng = Rng::new(peripherals.RNG);
 
@@ -238,7 +239,7 @@ async fn main_fallible(
             rst: peripherals.GPIO10,
             dc: peripherals.GPIO19,
             spi2: peripherals.SPI2,
-            dma: peripherals.DMA,
+            dma: peripherals.DMA_CH0,
         },
         history,
     )?;
@@ -254,7 +255,7 @@ async fn main_fallible(
         },
         clock.clone(),
         sender,
-    );
+    )?;
 
     info!("Stay awake for {}s", AWAKE_PERIOD.as_secs());
     Timer::after(AWAKE_PERIOD).await;
@@ -319,8 +320,8 @@ struct DisplayPeripherals {
     /// SPI interface
     spi2: SPI2,
 
-    /// DMA interface
-    dma: DMA,
+    /// DMA channel
+    dma: DmaChannel0,
 }
 
 /// Setup display task
@@ -330,15 +331,12 @@ fn setup_display_task(
     history: &'static mut HistoryBuffer<(OffsetDateTime, Sample), 96>,
 ) -> Result<Sender<'static, NoopRawMutex, (OffsetDateTime, Sample), 3>, Error> {
     info!("Create SPI bus");
-    let spi_config = SpiConfig {
-        frequency: 25_u32.kHz(),
-        mode: SpiMode::Mode0,
-        ..Default::default()
-    };
-    let spi_bus = Spi::new_with_config(peripherals.spi2, spi_config)
+    let spi_config = SpiConfig::default()
+        .with_frequency(25_u32.kHz())
+        .with_mode(SpiMode::_0);
+    let spi_bus = Spi::new(peripherals.spi2, spi_config)?
         .with_sck(peripherals.sclk)
-        .with_mosi(peripherals.mosi)
-        .into_async();
+        .with_mosi(peripherals.mosi);
 
     info!("Wrap SPI bus in a SPI DMA");
     let descriptors: &'static mut _ = DESCRIPTORS.init([DmaDescriptor::EMPTY; DESCRIPTORS_SIZE]);
@@ -348,10 +346,8 @@ fn setup_display_task(
     let buffer: &'static mut _ = BUFFER.init([0; BUFFERS_SIZE]);
     let rx_buffer: &'static mut _ = RX_BUFFER.init([0; BUFFERS_SIZE]);
 
-    let dma = Dma::new(peripherals.dma);
-    let dma_channel = dma.channel0.configure(false, DmaPriority::Priority0);
-
-    let spi_dma: SpiDma<'_, Async> = spi_bus.with_dma(dma_channel);
+    let spi_dma: SpiDma<'_, Blocking> = spi_bus.with_dma(peripherals.dma);
+    let spi_dma: SpiDma<'_, Async> = spi_dma.into_async();
 
     let tx_buffers = DmaTxBuf::new(descriptors, buffer)?;
     let rx_buffers = DmaRxBuf::new(rx_descriptors, rx_buffer)?;
@@ -402,13 +398,10 @@ fn setup_sensor_task(
     peripherals: SensorPeripherals,
     clock: Clock,
     sender: Sender<'static, NoopRawMutex, (OffsetDateTime, Sample), 3>,
-) {
+) -> Result<(), Error> {
     info!("Create I²C bus");
-    let i2c_config = I2cConfig {
-        frequency: 25_u32.kHz(),
-        ..Default::default()
-    };
-    let i2c = I2c::new(peripherals.i2c0, i2c_config)
+    let i2c_config = I2cConfig::default().with_frequency(25_u32.kHz());
+    let i2c = I2c::new(peripherals.i2c0, i2c_config)?
         .with_sda(peripherals.sda)
         .with_scl(peripherals.scl)
         .into_async();
@@ -420,6 +413,8 @@ fn setup_sensor_task(
         clock,
         SAMPLING_PERIOD,
     ));
+
+    Ok(())
 }
 
 /// An error
@@ -442,6 +437,13 @@ enum Error {
     /// An error within creation of DMA buffers
     #[expect(unused, reason = "Never read directly")]
     DmaBuffer(DmaBufError),
+
+    /// An error within creation of SPI bus
+    SpiConfig(SpiConfigError),
+
+    /// An error within creation of I²C bus
+    #[expect(unused, reason = "Never read directly")]
+    I2cConfig(I2cConfigError),
 }
 
 impl From<Infallible> for Error {
@@ -465,5 +467,17 @@ impl From<ClockError> for Error {
 impl From<DmaBufError> for Error {
     fn from(error: DmaBufError) -> Self {
         Self::DmaBuffer(error)
+    }
+}
+
+impl From<SpiConfigError> for Error {
+    fn from(error: SpiConfigError) -> Self {
+        Self::SpiConfig(error)
+    }
+}
+
+impl From<I2cConfigError> for Error {
+    fn from(error: I2cConfigError) -> Self {
+        Self::I2cConfig(error)
     }
 }